Add other pages and configure server

+++ src/main/java/com/site/plazam/domain/Rating.java
@@ -22,7 +22,7 @@ public class Rating {
 
     @NotNull
     @Field(name = "user_rating")
-    private Double userRating;
+    private Integer userRating;
 
     @NotNull
     @Field(name = "user_id")
@@ -32,7 +32,7 @@ public class Rating {
     @Field(name = "movie_id")
     private String movieId;
 
-    public Rating(@NotNull Double userRating,
+    public Rating(@NotNull Integer userRating,
                   @NotNull String userId,
                   @NotNull String movieId) {
         this.userRating = userRating;
